Extract and improve default logic forging :gear:
